EnumerationData::query_condition_set_members()
