+++ tiledb/api/src/query/strategy/tests.rs
@@ -11,9 +11,7 @@ use tiledb_common::query::condition::{
 };
 use tiledb_common::range::{NonEmptyDomain, Range, VarValueRange};
 use tiledb_common::Datatype;
-use tiledb_pod::array::schema::{
-    EnumerationData, FieldData as SchemaField, SchemaData,
-};
+use tiledb_pod::array::schema::{FieldData as SchemaField, SchemaData};
 use uri::TestArrayUri;
 
 use super::*;
@@ -646,54 +644,12 @@ impl QueryConditionField for FieldWithDomain {
         match self.field {
             SchemaField::Dimension(_) => None,
             SchemaField::Attribute(ref a) => {
-                let Some(edata) = self
+                let edata = self
                     .schema
-                    .enumeration(EnumerationKey::AttributeName(&a.name))
-                else {
-                    return None;
-                };
-                enumeration_records_to_set_members(edata)
+                    .enumeration(EnumerationKey::AttributeName(&a.name))?;
+                edata.query_condition_set_members()
             }
         }
-    }
-}
-
-fn enumeration_records_to_set_members(
-    edata: &EnumerationData,
-) -> Option<SetMembers> {
-    let records = edata.records();
-    if matches!(edata.datatype, Datatype::StringAscii | Datatype::StringUtf8)
-        && !matches!(edata.cell_val_num, Some(CellValNum::Fixed(_)))
-    {
-        Some(
-            records
-                .into_iter()
-                .map(|v| String::from_utf8_lossy(v.as_slice()).into_owned())
-                .collect::<Vec<String>>()
-                .into(),
-        )
-    } else if edata
-        .cell_val_num
-        .map(|c| c.is_single_valued())
-        .unwrap_or(true)
-    {
-        physical_type_go!(edata.datatype, DT, {
-            const WIDTH: usize = std::mem::size_of::<DT>();
-            type ByteArray = [u8; WIDTH];
-            Some(SetMembers::from(
-                records
-                    .into_iter()
-                    .map(|v| {
-                        assert_eq!(WIDTH, v.len());
-                        DT::from_le_bytes(
-                            ByteArray::try_from(v.as_slice()).unwrap(),
-                        )
-                    })
-                    .collect::<Vec<_>>(),
-            ))
-        })
-    } else {
-        None
     }
 }
